refactor(server): extract lobby manager
- move lobby caching, cleanup, and socket handlers into a
  dedicated component while preserving existing websocket behavior
